+++ spidermon/contrib/actions/slack/__init__.py
@@ -4,6 +4,7 @@ import logging
 
 from slack_sdk import WebClient
 from slack_sdk.errors import SlackApiError
+from slack_sdk.http_retry.builtin_handlers import RateLimitErrorRetryHandler
 
 from spidermon.contrib.actions.templates import ActionWithTemplates
 from spidermon.exceptions import NotConfigured
@@ -30,6 +31,7 @@ class SlackMessageManager:
 
         self.fake = fake
         self._client = WebClient(sender_token)
+        self._client.retry_handlers.append(RateLimitErrorRetryHandler())
         self._users = None
 
     @property
@@ -38,7 +40,7 @@ class SlackMessageManager:
             self._users = self._get_users_info()
         return self._users
 
-    def send_message(  # noqa: PLR0913
+    def send_message(  # noqa: PLR0913, PLR0917
         self,
         to,
         text,
@@ -160,7 +162,7 @@ class SlackMessageManager:
                 # can be an expected outcome - will use its own icon
                 icon_url = None
             else:
-                raise e
+                raise
         except KeyError:
             # bot has read permissions for slack org but can't find sender in list
             # can be an expected outcome - will use its own icon
@@ -189,7 +191,7 @@ class SendSlackMessage(ActionWithTemplates):
     include_attachments = True
     fake = False
 
-    def __init__(  # noqa: PLR0913
+    def __init__(  # noqa: PLR0913, PLR0917
         self,
         sender_token=None,
         sender_name=None,