+++ spidermon/contrib/validation/jsonschema/tools.py
@@ -14,14 +14,14 @@ def get_schema_from(source):
         schema = get_contents(source)
         try:
             return json.loads(schema)
-        except Exception as e:
-            logger.exception(str(e) + f"\nCould not parse schema from '{source}'")
+        except Exception:
+            logger.exception(f"Could not parse schema from '{source}'")
     elif source.endswith(".json"):
         with Path(source).open() as f:
             try:
                 return json.load(f)
-            except Exception as e:
-                logger.exception(str(e) + f"\nCould not parse schema in '{source}'")
+            except Exception:
+                logger.exception(f"Could not parse schema in '{source}'")
     else:
         schema = load_object(source)
         if isinstance(schema, str):
@@ -41,5 +41,5 @@ def get_contents(url):
     try:
         with urlopen(url) as f:  # noqa: S310
             return f.read().decode("utf-8")
-    except Exception as e:
-        logger.exception(str(e) + f"\nFailed to get '{url}'")
+    except Exception:
+        logger.exception(f"Failed to get '{url}'")