+++ spidermon/core/suites.py
@@ -25,7 +25,7 @@ class MonitorSuite(TestSuite, metaclass=MonitorOptionsMetaclass):
     monitors_passed_actions: ClassVar[list[str]] = []
     monitors_failed_actions: ClassVar[list[str]] = []
 
-    def __init__(  # noqa: PLR0913
+    def __init__(  # noqa: PLR0913, PLR0917
         self,
         name=None,
         monitors=None,