+++ tests/contrib/validation/jsonschema/test_tools.py
@@ -51,6 +51,6 @@ def test_get_contents_fails(mocker, caplog):
         (
             "spidermon.contrib.validation.jsonschema.tools",
             40,
-            "'ValueError' object has no attribute 'decode'\nFailed to get 'https://example.com/schema.json'",
+            "Failed to get 'https://example.com/schema.json'",
         ),
     ]